fix: fix authentication problem
BREAKING CHANGE:

+++ bt2/tests/article/test_delete_article.py
@@ -1,5 +1,6 @@
 import werkzeug
 
+from api.common.util import get_auth_token_for_test
 from api.db import get_db
 from api.service.article_service import insert_one_article
 import shortuuid
@@ -15,16 +16,19 @@ def test_delete_exist_article(test_client):
         "link": random_link, "title": random_title, "content": "fssdfsfsdfc"
     }
     result = jsonify(insert_one_article(db, data)).get_json()["id"]
-    print(result)
-    response = test_client.delete(f"api/articles/{int(result)}")
+    auth_token = get_auth_token_for_test(test_client)
+    response = test_client.delete(f"api/articles/{int(result)}",headers={"Authorization": "Bearer " +auth_token }, )
     assert response.status_code == 200
 
 
 def test_delete_non_exist_article(test_client):
-    response = test_client.delete(f"api/articles/1000000000")
+    auth_token = get_auth_token_for_test(test_client)
+    response = test_client.delete(f"api/articles/1000000000", headers={"Authorization": "Bearer " + auth_token })
     assert response.status_code == 404
 
 
 def test_delete_without_article_id(test_client):
-    response = test_client.delete(f"api/articles/")
+    auth_token = get_auth_token_for_test(test_client)
+
+    response = test_client.delete(f"api/articles/", headers={"Authorization": "Bearer " + auth_token })
     assert response.status_code == 405
